Correção na estrutura da store de auth
