+++ src/environments/environments.ts
@@ -1,4 +1,4 @@
 export const environment = {
-  apiUrl: "http://localhost:8000/api",
+  apiUrl: "https://api-hml-rancharia.atenaagencia.com/api",
 };
