+++ src/store/authStore.ts
@@ -1,7 +1,7 @@
 import { defineStore } from "pinia";
 import { ref } from "vue";
 import { ILogin } from "@/interfaces/loginInterface";
-import { login } from "../apis/authApi";
+import { login, logout } from "../apis/authApi";
 
 export const useAuthStore = defineStore("auth-store", () => {
   const token = ref<string | null>(null);
@@ -11,13 +11,21 @@ export const useAuthStore = defineStore("auth-store", () => {
       const response = await login(formLoginData);
       token.value = `${response.token_type} ${response.access_token}`;
       localStorage.setItem("token", token.value);
-      console.log(token.value);
       return response;
     } catch (error: any) {
-      console.error("Erro ao fazer login:", error.message);
       throw error;
     }
   };
 
-  return { token, userLogin };
+ const userLogout = async () => {
+       try{
+        const response = await logout()
+        return response
+       }
+       catch (error: any) {
+        return error.response?.data || error
+       }
+    }
+    
+  return { userLogin, userLogout };
 });
